Screenshot of example folder
Here you see a few files that carbon removal equipment might create as well as XML files that can contain things like the link to the PDF of a lab that has checked the carbon content. The data in that XML file correspond to the date in the PDF and this allows software that follows a strict methodology to ensure it hasn't been longer than the required duration of time since the lab result as stated by that software auditors methodology. These methodologies can be in the smart contract itself to ensure only real carbon removal that follows all rules will trigger payment when fully delivered. You also see a total.php which shows the concept of a script that can total everything in a series of folders by just applying math to the meta data. And the idea of a deduction for margins of error as measured by a third party visitation is reflected. This .XML file can include the ratio of that deduction as well as the link to that report provided by the company who performed it. If that company is out of business and is no longer standing behind that 3rd party audit and so isn't hosting the file anymore then that tells you something right there. But at least while they exist it allows a piece of software performing real time auditing per a methodology to ensure all stated rules are followed. Some files that might make less sense at first glance would be: 

sinking_speed_data_vessel_01.xml
load_cell_data_vessel_01.xml

These are example files that would have raw data. While a company may feel the way they turn their raw data into an inferred total value could be private and part of their secret sauce the raw data itself should be made available to the public so that others can reproduce similar inferring systems to make sure they can reproduce the calculations. Otherwise a company can't really be trusted to conduct their own measurements. This idea reflects the need to balance investors need for startups to have some closed IP with the academic communities need to have ensure accountability of carbon removal solutions that are catching on and should be closely scrutinized.

And finally all the folders are just meant to show years, months, days and batches. The name of a folder containing the data proving carbon was removed can have a lot of data. In the shown example you see lat and long data, time of day in UTC down to the minute and KG. This should be before the ratio in the .XML is applied. In this case that would be wet weight of sinking seaweed then with a lab provided ratio of actual emissions of CO2 reversed. Those ratios are determined by the lab measuring the wet weight of what they receive and then performing dry mass analysis to get the molecular weight of the constituent parts. Then the scientifically known ratio of carbon to CO2 (in which the oxygen would be present) is applied. 

Quick note for those not closely following the biology, chemistry and the carbon removal industry: Photosynthesis removes the oxygen which accounts for a lot of the mass of CO2 in the atmosphere. And the carbon removal industry counts emissions reversed and so this math must be applied to represent the measured carbon content as carbon removed.